Enable a web-like structure for the schema.
A goal may be part of more than one objective.
A task may be part of more than one goal.

Wrap the creation of the schema in a stored procedure.

This will allow for the schema to be created in a single call.
It will also allow the data to be retained between
schema changes in a more consistent manner.