Track users work activity in the work_log entity
Enhance the interface to the show the user whether
they are currently working a task or not. Also,
show them how much work they have accomplished.

Ensure only one task is running per user at any
given time.

If a goal is added to a completed objective then
set the objective's completed_dtm value to null.
If a task is added to a completed goal then
set the goal's completed_dtm value to null.